// src/games/stack-em/types/Card.ts
export type Suit = 'hearts' | 'diamonds' | 'clubs' | 'spades';
export type Rank = 'A' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | 'J' | 'Q' | 'K';

export type Card = {
  suit: Suit;
  rank: Rank;
  value: number;
  isDealt: boolean;
  isSwapCard: boolean;
  isWildCard: boolean;
};

export type Deck = Card[];